requirements-dev.txt
-r requirements.txt

ipdb
mock
django-debug-toolbar
django-test-without-migrations
django-bootstrap-form
